[zen-27] Renamed interfaces in Zend\Soap
