+++ library/Zend/Soap/Client/Local.php
@@ -32,8 +32,6 @@ use Zend\Soap\Client as SOAPClient,
  *
  * Could be used for development or testing purposes.
  *
- * @uses       \Zend\Soap\Client
- * @uses       \Zend\Soap\Server
  * @category   Zend
  * @package    Zend_Soap
  * @subpackage Client
