+++ library/Zend/Soap/Exception/BadMethodCallException.php
@@ -4,6 +4,6 @@ namespace Zend\Soap\Exception;
 
 class BadMethodCallException
     extends \BadMethodCallException
-    implements \Zend\Soap\Exception
+    implements ExceptionInterface
 {
 }
