+++ library/Zend/Soap/Exception/ExceptionInterface.php
@@ -14,21 +14,20 @@
  *
  * @category   Zend
  * @package    Zend_Soap
- * @subpackage WSDL
+ * @subpackage AutoDiscover
  * @copyright  Copyright (c) 2005-2012 Zend Technologies USA Inc. (http://www.zend.com)
  * @license    http://framework.zend.com/license/new-bsd     New BSD License
  */
 
-namespace Zend\Soap;
+namespace Zend\Soap\Exception;
 
 /**
- * Zend_Soap_Wsdl_Exception
- *
- * @uses       \Zend\Soap\Exception
  * @category   Zend
  * @package    Zend_Soap
- * @subpackage Wsdl
+ * @subpackage AutoDiscover
  * @copyright  Copyright (c) 2005-2012 Zend Technologies USA Inc. (http://www.zend.com)
  * @license    http://framework.zend.com/license/new-bsd     New BSD License
  */
-class WsdlException extends Exception { }
+interface ExceptionInterface
+{
+}
