+++ library/Zend/Soap/Exception/ExceptionNotLoadedException.php
@@ -4,6 +4,6 @@ namespace Zend\Soap\Exception;
 
 class ExtensionNotLoadedException
     extends \RuntimeException
-    implements \Zend\Soap\Exception
+    implements ExceptionInterface
 {
 }
