+++ library/Zend/Soap/Exception/InvalidArgumentException.php
@@ -4,6 +4,6 @@ namespace Zend\Soap\Exception;
 
 class InvalidArgumentException
     extends \InvalidArgumentException
-    implements \Zend\Soap\Exception
+    implements ExceptionInterface
 {
 }
