+++ library/Zend/Soap/Exception/RuntimeException.php
@@ -4,7 +4,7 @@ namespace Zend\Soap\Exception;
 
 class RuntimeException
     extends \RuntimeException
-    implements \Zend\Soap\Exception
+    implements ExceptionInterface
 {
     
 }
