+++ library/Zend/Soap/Exception/UnexpectedValueException.php
@@ -4,7 +4,7 @@ namespace Zend\Soap\Exception;
 
 class UnexpectedValueException
     extends \UnexpectedValueException
-    implements \Zend\Soap\Exception
+    implements ExceptionInterface
 {
     
 }
