+++ library/Zend/Soap/Wsdl/ComplexTypeStrategy/AnyType.php
@@ -21,19 +21,16 @@
 
 namespace Zend\Soap\Wsdl\ComplexTypeStrategy;
 
-use Zend\Soap\Wsdl\ComplexTypeStrategy;
-
 /**
  * Zend_Soap_Wsdl_Strategy_AnyType
  *
- * @uses       \Zend\Soap\Wsdl\Strategy\StrategyInterface
  * @category   Zend
  * @package    Zend_Soap
  * @subpackage WSDL
  * @copyright  Copyright (c) 2005-2012 Zend Technologies USA Inc. (http://www.zend.com)
  * @license    http://framework.zend.com/license/new-bsd     New BSD License
  */
-class AnyType implements ComplexTypeStrategy
+class AnyType implements ComplexTypeStrategyInterface
 {
     /**
      * Not needed in this strategy.
