+++ library/Zend/Soap/Wsdl/ComplexTypeStrategy/ArrayOfTypeComplex.php
@@ -21,16 +21,13 @@
 
 namespace Zend\Soap\Wsdl\ComplexTypeStrategy;
 
-use Zend\Soap;
-
-use Zend\Soap\Wsdl;
-use Zend\Soap\Exception;
+use Zend\Soap,
+    Zend\Soap\Wsdl,
+    Zend\Soap\Exception;
 
 /**
  * ArrayOfTypeComplex strategy
  *
- * @uses       \Zend\Soap\Wsdl\Exception
- * @uses       \Zend\Soap\Wsdl\Strategy\DefaultComplexType
  * @category   Zend
  * @package    Zend_Soap
  * @subpackage WSDL
