+++ tests/Zend/Soap/ServerTest.php
@@ -32,7 +32,6 @@ use Zend\Soap\Server,
  * @category   Zend
  * @package    Zend_Soap
  * @subpackage UnitTests
- * @uses       Zend_Server_Interface
  * @copyright  Copyright (c) 2005-2012 Zend Technologies USA Inc. (http://www.zend.com)
  * @license    http://framework.zend.com/license/new-bsd     New BSD License
  * @group      Zend_Soap
