+++ tests/Zend/Soap/Wsdl/CompositeStrategyTest.php
@@ -19,9 +19,9 @@
  */
 
 namespace ZendTest\Soap\Wsdl;
+
 use Zend\Soap\Wsdl\ComplexTypeStrategy,
     Zend\Soap\Wsdl,
-    Zend\Soap\WsdlException,
     Zend\Soap\Wsdl\ComplexTypeStrategy\Composite,
     Zend\Soap\Wsdl\ComplexTypeStrategy\ArrayOfTypeComplex,
     Zend\Soap\Wsdl\ComplexTypeStrategy\ArrayOfTypeSequence;
